+++ src/main/java/es/swapsounds/repository/SoundRepository.java
@@ -3,7 +3,7 @@ package es.swapsounds.repository;
 import es.swapsounds.model.Sound;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SoundRepository extends JpaRepository <Sound, Long> {
+public interface SoundRepository extends JpaRepository <Sound, Integer> {
 
     
 }
